Move AccountUnitTest inside AccountCreationActivityAndAccountTest
This has been done in order to use the activity context
