+++ app/src/main/java/ch/epfl/sweng/SDP/localDatabase/LocalDbHandlerForAccount.java
@@ -18,6 +18,10 @@ public class LocalDbHandlerForAccount extends SQLiteOpenHelper {
     private static final String COLUMN_CURRENT_LEAGUE = "currentLeague";
     private static final String COLUMN_TROPHIES = "trophies";
     private static final String COLUMN_STARS = "stars";
+    private static final String COLUMN_MATCHES_WON = "matchesWon";
+    private static final String COLUMN_MATCHES_LOST = "matchesLost";
+    private static final String COLUMN_AVERAGE_RATING = "averageRating";
+    private static final String COLUMN_MAX_TROPHIES = "maxTrophies";
 
     /**
      * Helper class to save the account in local database.
@@ -37,7 +41,9 @@ public class LocalDbHandlerForAccount extends SQLiteOpenHelper {
         String createTable = "CREATE TABLE " + TABLE_NAME + "(" + COLUMN_ID
                 + " INTEGER PRIMARY KEY," + COLUMN_USER_ID + " USER_ID," + COLUMN_USERNAME
                 + " USERNAME," + COLUMN_CURRENT_LEAGUE + " CURRENT_LEAGUE," + COLUMN_TROPHIES
-                + " TROPHIES," + COLUMN_STARS + " STARS )";
+                + " TROPHIES," + COLUMN_STARS + " STARS," + COLUMN_MATCHES_WON + " MATCHES_WON,"
+                + COLUMN_MATCHES_LOST + " MATCHES_LOST," + COLUMN_AVERAGE_RATING
+                + " AVERAGE_RATING," + COLUMN_MAX_TROPHIES + " MAX_TROPHIES )";
 
         db.execSQL(createTable);
     }
@@ -67,6 +73,10 @@ public class LocalDbHandlerForAccount extends SQLiteOpenHelper {
         values.put(COLUMN_CURRENT_LEAGUE, account.getCurrentLeague());
         values.put(COLUMN_TROPHIES, account.getTrophies());
         values.put(COLUMN_STARS, account.getStars());
+        values.put(COLUMN_MATCHES_WON, account.getMatchesWon());
+        values.put(COLUMN_MATCHES_LOST, account.getMatchesLost());
+        values.put(COLUMN_AVERAGE_RATING, account.getAverageRating());
+        values.put(COLUMN_MAX_TROPHIES, account.getMaxTrophies());
 
         SQLiteDatabase db = this.getWritableDatabase();
         db.insert(TABLE_NAME, null, values);
@@ -93,9 +103,14 @@ public class LocalDbHandlerForAccount extends SQLiteOpenHelper {
             account.setCurrentLeague(cursor.getString(3));
             account.setTrophies(cursor.getInt(4));
             account.setStars(cursor.getInt(5));
+            account.setMatchesWon(cursor.getInt(6));
+            account.setMatchesLost(cursor.getInt(7));
+            account.setAverageRating(cursor.getDouble(8));
+            account.setMaxTrophies(cursor.getInt(9));
             account.setUsersRef(Database.INSTANCE.getReference("users"));
             cursor.close();
         }
         db.close();
     }
 }
+
